requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
Brotli==1.0.9
cffi==1.15.0
colorama==0.4.3
contourpy==1.2.0
cssselect2==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20230430
distlib==0.3.1
Django==4.2.9
django-crispy-forms==1.10.0
django-extensions==3.1.5
django-filter==21.1
django-import-export==3.3.6
django-qr-code==2.2.0
entrypoints==0.3
et_xmlfile==1.0.1
filelock==3.0.12
fonttools==4.27.1
html5lib==1.1
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.4.5
lml==0.1.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.8.2
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.3
pandas==2.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prettytable==0.7.2
prometheus_client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
pydyf==0.1.1
pyexcel-io==0.6.4
pyexcel-xls==0.6.1
pyexcel-xlsx==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyphen==0.11.0
pyrsistent==0.16.0
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.3.1
scipy==1.11.4
seaborn==0.10.1
segno==1.3.3
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
tablib==3.5.0
terminado==0.8.3
testpath==0.4.4
tinycss2==1.1.0
tornado==6.0.4
traitlets==4.3.3
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.2.2
wcwidth==0.1.9
weasyprint==53.3
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
zopfli==0.1.9
